+++ src/test/java/org/realityforge/bazel/depgen/record/ApplicationRecordTest.java
@@ -3473,7 +3473,7 @@ public class ApplicationRecordTest
                   "\n" +
                   "load(\"@bazel_tools//tools/build_defs/repo:http.bzl\", _http_file = \"http_file\")\n" +
                   "load(\"@rules_java//java:defs.bzl\", _java_binary = \"java_binary\", _java_import = \"java_import\", _java_test = \"java_test\")\n" +
-                  "load(\"@com_google_j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
+                  "load(\"@j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
                   "\n" +
                   "# SHA256 of the configuration content that generated this file\n" +
                   "_CONFIG_SHA256 = \"MYSHA\"\n" +
@@ -3601,7 +3601,7 @@ public class ApplicationRecordTest
                   "\n" +
                   "load(\"@bazel_tools//tools/build_defs/repo:http.bzl\", _http_file = \"http_file\")\n" +
                   "load(\"@rules_java//java:defs.bzl\", _java_binary = \"java_binary\", _java_import = \"java_import\", _java_test = \"java_test\")\n" +
-                  "load(\"@com_google_j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
+                  "load(\"@j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
                   "\n" +
                   "# SHA256 of the configuration content that generated this file\n" +
                   "_CONFIG_SHA256 = \"MYSHA\"\n" +
@@ -3741,7 +3741,7 @@ public class ApplicationRecordTest
                   "\n" +
                   "load(\"@bazel_tools//tools/build_defs/repo:http.bzl\", _http_file = \"http_file\", _http_archive = \"http_archive\")\n" +
                   "load(\"@rules_java//java:defs.bzl\", _java_binary = \"java_binary\", _java_import = \"java_import\", _java_test = \"java_test\")\n" +
-                  "load(\"@com_google_j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
+                  "load(\"@j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
                   "\n" +
                   "# SHA256 of the configuration content that generated this file\n" +
                   "_CONFIG_SHA256 = \"MYSHA\"\n" +
@@ -4202,7 +4202,7 @@ public class ApplicationRecordTest
                   "\n" +
                   "load(\"@bazel_tools//tools/build_defs/repo:http.bzl\", _http_file = \"http_file\")\n" +
                   "load(\"@rules_java//java:defs.bzl\", _java_binary = \"java_binary\", _java_import = \"java_import\", _java_library = \"java_library\", _java_plugin = \"java_plugin\", _java_test = \"java_test\")\n" +
-                  "load(\"@com_google_j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
+                  "load(\"@j2cl//build_defs:rules.bzl\", _j2cl_library = \"j2cl_library\")\n" +
                   "\n" +
                   "# SHA256 of the configuration content that generated this file\n" +
                   "_CONFIG_SHA256 = \"MYSHA\"\n" +